working on test passage
