+++ lux/extensions/content/models.py
@@ -1,13 +1,9 @@
 import os
-import mimetypes
 
-from pulsar import Http404
 from pulsar.utils.httpurl import remove_double_slash
-from pulsar.utils.slugify import slugify
 
-from lux.core import cached, models
-from lux.extensions.rest import RestModel, RestField
-from lux.utils.data import as_tuple
+from lux.core import cached
+from lux.extensions.rest import RestModel, RestField, Query
 from lux.utils.files import skipfile
 
 from .contents import get_reader
@@ -18,18 +14,7 @@ FIELDS = [
     RestField('order', sortable=True, type='int'),
     RestField('slug', sortable=True),
     RestField('group', sortable=True),
-    RestField('path', sortable=True),
     RestField('title')]
-
-
-OPERATORS = {
-    'eq': lambda x, y: x == y,
-    'ne': lambda x, y: x != y,
-    'gt': lambda x, y: x > y,
-    'ge': lambda x, y: x >= y,
-    'lt': lambda x, y: x < y,
-    'le': lambda x, y: x <= y
-}
 
 
 class ContentModel(RestModel):
@@ -43,7 +28,7 @@ class ContentModel(RestModel):
         self.directory = location
         self.ext = ext
         fields = fields or FIELDS[:]
-        kw['id_field'] = 'path'
+        kw['id_field'] = 'slug'
         super().__init__(name, fields=fields, **kw)
 
     def session(self, request, session=None):
@@ -77,28 +62,8 @@ class ContentModel(RestModel):
         return body
 
 
-class QuerySession(models.Query):
-    _filtered_data = None
-    _data = None
-    _limit = None
-    _offset = None
-    _paths = None
+class Session:
 
-    def __init__(self, model, request):
-        super().__init__(model)
-        self._filters = []
-        self._groups = []
-        self._sortby = []
-        self.request = request
-
-    def __repr__(self):
-        if self._data is None:
-            return self.__class__.__name__
-        else:
-            return repr(self._data)
-    __str__ = __repr__
-
-    # Session methods
     def __enter__(self):
         return self
 
@@ -108,62 +73,18 @@ class QuerySession(models.Query):
     def add(self, instance):
         pass
 
+    def delete(self, instance):
+        pass
+
     def flush(self):
         pass
 
-    # Query methods
-    def one(self):
-        data = self.all()
-        if data:
-            if len(data) > 1:
-                self.request.logger.error('Multiple result found for model %s.'
-                                          'returning the first' % self.name)
-            return data[0]
-        else:
-            raise Http404
 
-    def limit(self, v):
-        self._limit = v
-        return self
+class QuerySession(Query, Session):
 
-    def offset(self, v):
-        self._offset = v
-        return self
-
-    def count(self):
-        return len(self._get_data())
-
-    def filter_args(self, args):
-        self._filtered_data = None
-        self._paths = args
-
-    def filter_field(self, field, op, value):
-        self._filtered_data = None
-        if op in OPERATORS:
-            op = OPERATORS[op]
-            if field == 'group':
-                self._groups.extend(as_tuple(value))
-            self._filters.append((field, op, value))
-        else:
-            self.app.logger.error('Could not apply filter %s to %s',
-                                  op, self)
-
-    def sortby_field(self, field, direction):
-        self._sortby.append((field, direction))
-
-    def all(self):
-        data = self._get_data()
-        for field, direction in self._sortby:
-            if direction == 'desc':
-                data = [desc(d, field) for d in data]
-            else:
-                data = [asc(d, field) for d in data]
-            data = [s.d for s in sorted(data)]
-        if self._offset:
-            data = data[self._offset:]
-        if self._limit:
-            data = data[:self._limit]
-        return data
+    def __init__(self, model, request):
+        super().__init__(model, request)
+        self._groups = []
 
     #  INTERNALS
     def _get_data(self):
@@ -179,10 +100,6 @@ class QuerySession(models.Query):
                 if self._filter(content):
                     data.append(model.instance(content, self.fields))
         return self._filtered_data
-
-    def _sort(self, c):
-        if self._sort_field in c:
-            return
 
     def _filter(self, content):
         for field, op, value in self._filters:
@@ -219,49 +136,3 @@ class QuerySession(models.Query):
                                 slug=slug)
                     data.append(reader.read(src, meta).tojson())
         return data
-
-
-class asc:
-    __slots__ = ('d', 'value')
-
-    def __init__(self, d, field):
-        self.d = d
-        self.value = d.get(field)
-
-    def __eq__(self, other):
-        return self.value == other.value
-
-    def __lt__(self, other):
-        if self.value is None:
-            return False
-        elif other.value is None:
-            return True
-        else:
-            return self.value < other.value
-
-    def __gt__(self, other):
-        if other.value is None:
-            return False
-        elif self.value is None:
-            return True
-        else:
-            return self.value > other.value
-
-
-class desc(asc):
-
-    def __gt__(self, other):
-        if self.value is None:
-            return False
-        elif other.value is None:
-            return True
-        else:
-            return self.value < other.value
-
-    def __lt__(self, other):
-        if self.value is None:
-            return False
-        elif other.value is None:
-            return True
-        else:
-            return self.value > other.value
