+++ lux/utils/test.py
@@ -129,6 +129,9 @@ def create_users(app, items):
 
 
 async def load_fixtures(app, path=None):
+    if not hasattr(app.auth_backend, 'create_user'):
+        return
+
     fpath = path if path else os.path.join(app.meta.path, 'fixtures')
 
     fixtures = OrderedDict()
@@ -153,7 +156,7 @@ async def load_fixtures(app, path=None):
     test_token = test.json(request.response, 201)['token']
 
     for model, items in fixtures.items():
-        logger.debug('Creating %d fixtures for "%s"', len(items), model)
+        logger.info('Creating %d fixtures for "%s"', len(items), model)
         for params in items:
             url = params.pop('api_url', '/%s' % model)
             request = await client.post(url,
@@ -558,6 +561,22 @@ class AppTestCase(unittest.TestCase, TestMixin):
                                        ['--username', username,
                                         '--email', email,
                                         '--password', password])
+
+    async def _token(self, credentials):
+        '''Return a token for a user
+        '''
+        if isinstance(credentials, str):
+            credentials = {"username": credentials,
+                           "password": credentials}
+
+        # Get new token
+        request = await self.client.post('/authorizations',
+                                         json=credentials)
+        user = request.cache.user
+        self.assertFalse(user.is_authenticated())
+        data = self.json(request.response, 201)
+        self.assertTrue('token' in data)
+        return data['token']
 
 
 class WebApiTestCase(AppTestCase):
