--- tests/fixtures/permissions.json
@@ -1,16 +1,0 @@
-{
-    "admin-read": {
-        "description": "Read only permission for the Admin site",
-        "policy": {
-            "resource": "site:admin",
-            "action": "read"
-        }
-    },
-    "add-document": {
-        "description": "Read add and update of documents",
-        "policy": {
-            "resource": "document",
-            "action": ["read", "create", "update"]
-        }
-    }
-}
