+++ tests/odm/test_postgresql.py
@@ -2,93 +2,10 @@ from dateutil.parser import parse
 
 from lux.utils import test
 
-
-class TestPostgreSqlBase(test.AppTestCase):
-    config_file = 'tests.odm'
-    config_params = {
-        'DATASTORE': 'postgresql+green://lux:luxtest@127.0.0.1:5432/luxtests'}
-    su_credentials = {'username': 'pippo',
-                      'password': 'pluto'}
-
-    @classmethod
-    def populatedb(cls):
-        backend = cls.app.auth_backend
-        backend.create_superuser(cls.app.wsgi_request(),
-                                 email='pippo@pippo.com',
-                                 first_name='Pippo',
-                                 **cls.su_credentials)
-
-    async def _token(self, credentials=None):
-        '''Create an authentication token
-        '''
-        if credentials is None:
-            credentials = self.su_credentials
-        request = await self.client.post('/authorizations', json=credentials)
-        response = request.response
-        self.assertEqual(response.status_code, 201)
-        user = request.cache.user
-        self.assertFalse(user.is_authenticated())
-        data = self.json(response)
-        self.assertTrue('token' in data)
-        return data['token']
-
-    async def _create_task(self, token, subject='This is a task', person=None,
-                           **data):
-        data['subject'] = subject
-        if person:
-            data['assigned'] = person['id']
-        request = await self.client.post('/tasks', json=data, token=token)
-        data = self.json(request.response, 201)
-        self.assertIsInstance(data, dict)
-        self.assertTrue('id' in data)
-        self.assertEqual(data['subject'], subject)
-        self.assertTrue('created' in data)
-        return data
-
-    async def _get_task(self, token, id):
-        request = await self.client.get(
-            '/tasks/{}'.format(id),
-            token=token)
-        response = request.response
-        self.assertEqual(response.status_code, 200)
-        data = self.json(response)
-        self.assertIsInstance(data, dict)
-        self.assertTrue('id' in data)
-        return data
-
-    async def _delete_task(self, token, id):
-        request = await self.client.delete(
-            '/tasks/{}'.format(id),
-            token=token)
-        response = request.response
-        self.assertEqual(response.status_code, 204)
-
-    async def _create_person(self, token, username, name=None):
-        name = name or username
-        request = await self.client.post(
-            '/people',
-            json={'username': username, 'name': name},
-            token=token)
-        data = self.json(request.response, 201)
-        self.assertIsInstance(data, dict)
-        self.assertTrue('id' in data)
-        self.assertEqual(data['name'], name)
-        return data
-
-    async def _update_person(self, token, id, username=None, name=None):
-        request = await self.client.post(
-            '/people/{}'.format(id),
-            json={'username': username, 'name': name},
-            token=token)
-        data = self.json(request.response, 200)
-        self.assertIsInstance(data, dict)
-        self.assertTrue('id' in data)
-        if name:
-            self.assertEqual(data['name'], name)
-        return data
+from tests.odm.utils import OdmUtils
 
 
-class TestPostgreSql(TestPostgreSqlBase):
+class TestPostgreSql(OdmUtils, test.AppTestCase):
 
     async def test_odm(self):
         tables = await self.app.odm.tables()
@@ -153,14 +70,14 @@ class TestPostgreSql(TestPostgreSqlBase):
         self.assertIsInstance(data, dict)
         columns = data['columns']
         self.assertIsInstance(columns, list)
-        self.assertEqual(len(columns), 7)
+        self.assertEqual(len(columns), 10)
 
     async def test_create_task(self):
-        token = await self._token()
+        token = await self._token('testuser')
         await self._create_task(token)
 
     async def test_update_task(self):
-        token = await self._token()
+        token = await self._token('testuser')
         task = await self._create_task(token, 'This is another task')
         # Update task
         request = await self.client.post(
@@ -189,7 +106,7 @@ class TestPostgreSql(TestPostgreSqlBase):
         self.assertEqual(data['done'], True)
 
     async def test_delete_task(self):
-        token = await self._token()
+        token = await self._token('testuser')
         task = await self._create_task(token, 'A task to be deleted')
         # Delete task
         request = await self.client.delete('/tasks/%d' % task['id'])
@@ -206,7 +123,7 @@ class TestPostgreSql(TestPostgreSqlBase):
         self.assertEqual(response.status_code, 404)
 
     async def test_sortby(self):
-        token = await self._token()
+        token = await self._token('testuser')
         await self._create_task(token, 'We want to sort 1')
         await self._create_task(token, 'We want to sort 2')
         request = await self.client.get('/tasks?sortby=created')
@@ -232,7 +149,7 @@ class TestPostgreSql(TestPostgreSqlBase):
             self.assertTrue(dt2 < dt1)
 
     async def test_sortby_non_existent(self):
-        token = await self._token()
+        token = await self._token('testuser')
         await self._create_task(token, 'a task')
         await self._create_task(token, 'another task')
         request = await self.client.get('/tasks?sortby=fgjsdgj')
@@ -241,7 +158,7 @@ class TestPostgreSql(TestPostgreSqlBase):
         self.assertIsInstance(result, list)
 
     async def test_relationship_field(self):
-        token = await self._token()
+        token = await self._token('testuser')
         person = await self._create_person(token, 'spiderman')
         task = await self._create_task(token,
                                        'climb a wall a day',
@@ -252,7 +169,7 @@ class TestPostgreSql(TestPostgreSqlBase):
         self.assertEqual(data['assigned'], task['assigned'])
 
     async def test_relationship_field_failed(self):
-        token = await self._token()
+        token = await self._token('testuser')
         data = {'subject': 'climb a wall a day',
                 'assigned': 6868897}
         request = await self.client.post('/tasks', json=data, token=token)
@@ -260,7 +177,7 @@ class TestPostgreSql(TestPostgreSqlBase):
                                    'Invalid person')
 
     async def test_unique_field(self):
-        token = await self._token()
+        token = await self._token('testuser')
         await self._create_person(token, 'spiderman1', 'luca')
         data = dict(username='spiderman1', name='john')
         request = await self.client.post('/people', json=data, token=token)
@@ -272,7 +189,7 @@ class TestPostgreSql(TestPostgreSqlBase):
         Tests that it's not possible to update a unique field of an
         existing record to that of another
         """
-        token = await self._token()
+        token = await self._token('testuser')
         await self._create_person(token, 'spiderfail1', 'luca')
         data = await self._create_person(token, 'spiderfail2', 'pippo')
 
@@ -289,20 +206,20 @@ class TestPostgreSql(TestPostgreSqlBase):
         Tests that an update of an existing model instance works if the the
         unique field hasn't changed
         """
-        token = await self._token()
+        token = await self._token('testuser')
         data = await self._create_person(token, 'spiderstale1', 'luca')
         await self._update_person(token, data['id'], 'spiderstale1',
                                   'lucachanged')
 
     async def test_enum_field(self):
-        token = await self._token()
+        token = await self._token('testuser')
         data = await self._create_task(token, enum_field='opt1')
         self.assertEqual(data['enum_field'], 'opt1')
         data = await self._get_task(token, id=data['id'])
         self.assertEqual(data['enum_field'], 'opt1')
 
     async def test_enum_field_fail(self):
-        token = await self._token()
+        token = await self._token('testuser')
         request = await self.client.post('/tasks', json={'enum_field': 'opt3'},
                                          token=token)
         response = request.response
@@ -321,7 +238,7 @@ class TestPostgreSql(TestPostgreSqlBase):
         self.assertEqual(response.status_code, 200)
         self.assertEqual(response['Access-Control-Allow-Methods'],
                          'GET, POST, DELETE')
-        token = await self._token()
+        token = await self._token('testuser')
         task = await self._create_task(token, 'testing preflight on id')
         request = await self.client.options('/tasks/%s' % task['id'])
         response = request.response
@@ -334,7 +251,7 @@ class TestPostgreSql(TestPostgreSqlBase):
         response = request.response
         self.assertEqual(response.status_code, 404)
         self.assertFalse(response.content)
-        token = await self._token()
+        token = await self._token('testuser')
         task = await self._create_task(token, 'testing head request')
         request = await self.client.head('/tasks/%s' % task['id'])
         response = request.response
@@ -342,7 +259,7 @@ class TestPostgreSql(TestPostgreSqlBase):
         self.assertFalse(response.content)
 
     async def test_filter(self):
-        token = await self._token()
+        token = await self._token('testuser')
         await self._create_task(token, 'A done task', done=True)
         await self._create_task(token, 'a not done task')
         request = await self.client.get('/tasks?done=1')
@@ -364,7 +281,7 @@ class TestPostgreSql(TestPostgreSqlBase):
             self.assertEqual(task['done'], False)
 
     async def test_limit(self):
-        token = await self._token()
+        token = await self._token('testuser')
         await self._create_task(token, 'whatever')
         await self._create_task(token, 'do everything')
         request = await self.client.get('/tasks?limit=-1')
