+++ tests/odm/test_sqlite.py
@@ -1,12 +1,7 @@
-from lux.utils import test
-
 import tests.odm.test_postgresql as postgresql
 
-
-class TestSqliteMixin:
-    config_params = {'DATASTORE': 'sqlite://'}
+from tests.odm.utils import SqliteMixin
 
 
-@test.sequential
-class TestSql(TestSqliteMixin, postgresql.TestPostgreSql):
+class TestSql(SqliteMixin, postgresql.TestPostgreSql):
     pass
