+++ tests/web/__init__.py
@@ -1,18 +1,11 @@
 from lux.utils import test
 
-from tests import load_fixture
-
-SUPERSTAFF = ('superman', 'spiderman', 'zorro')
-STAFF = ('pippo', 'pluto', 'toni')
+import tests
 
 
-class WebsiteTest(test.WebApiTestCase):
+class WebsiteTest(tests.AuthFixtureMixin, test.WebApiTestCase):
     config_file = 'example.webapi.config'
     web_config_file = 'example.website.config'
-
-    @classmethod
-    def populatedb(cls):
-        testdb(cls.app)
 
     async def _login(self, credentials=None, csrf=None, status=201):
         '''Return a token for a new superuser
@@ -27,7 +20,7 @@ class WebsiteTest(test.WebApiTestCase):
             self.assertTrue(cookie.startswith('test-website='))
         csrf = csrf or {}
         if credentials is None:
-            credentials = SUPERSTAFF[0]
+            credentials = 'testuser'
         if not isinstance(credentials, dict):
             credentials = dict(username=credentials,
                                password=credentials)
@@ -47,32 +40,3 @@ class WebsiteTest(test.WebApiTestCase):
             self.assertFalse(request.cache.user.is_authenticated())
             self.assertTrue('token' in data)
             return cookie2, data['token']
-
-
-def testdb(app):
-    odm = app.odm()
-    PERMISSIONS = load_fixture('permissions')
-    for engine in odm.engines():
-        app.logger.info('Load fixtures to %s', engine)
-
-    with odm.begin() as se:
-        for name, values in PERMISSIONS.items():
-            perm = odm.permission(name=name, **values)
-            se.add(perm)
-
-    # Add staff
-    backend = app.auth_backend
-    request = app.wsgi_request()
-    for staff in SUPERSTAFF:
-        backend.create_superuser(request,
-                                 username=staff,
-                                 email='%s@foo.com' % staff,
-                                 password=staff)
-    for staff in STAFF:
-        backend.create_user(request,
-                            username=staff,
-                            email='%s@foo.com' % staff,
-                            password=staff,
-                            active=True)
-
-    app.logger.info('Database fixtures loaded')
